requirements.txt
pillow
numpy
scikit-image
scikit-learn
einops
opencv-python
matplotlib
wandb
scipy
tqdm 
torchmetrics
fvcore
lightning
tensorboard